requirements.txt
opencv-python
opencv-python-headless
pillow
tk
tensorflow
numpy